+++ tests/helper.py
@@ -1,7 +1,11 @@
 import json
 
 from webdriver_manager.core.os_manager import ChromeType
-from webdriver_manager.drivers.chrome import ChromeDriver
+from webdriver_manager.drivers.chrome import (
+    CHROME_FOR_TESTING_DOWNLOAD_URL,
+    CHROME_FOR_TESTING_LATEST_RELEASE_URL,
+    ChromeDriver,
+)
 
 
 class ResponseMock:
@@ -33,15 +37,20 @@ class OperationSystemManagerMock:
         return self.browser_version
 
 
-def chrome_driver_for(browser_version, responses, chrome_type=ChromeType.CHROMIUM):
+def chrome_driver_for(
+        browser_version,
+        responses,
+        chrome_type=ChromeType.CHROMIUM,
+        driver_version=None,
+        url=CHROME_FOR_TESTING_DOWNLOAD_URL,
+        latest_release_url=CHROME_FOR_TESTING_LATEST_RELEASE_URL,
+):
     http_client = HttpClientMock(responses)
     driver = ChromeDriver(
         name="chromedriver",
-        driver_version=None,
-        url="https://storage.googleapis.com/chrome-for-testing-public/",
-        latest_release_url=(
-            "https://googlechromelabs.github.io/chrome-for-testing/LATEST_RELEASE_STABLE"
-        ),
+        driver_version=driver_version,
+        url=url,
+        latest_release_url=latest_release_url,
         http_client=http_client,
         os_system_manager=OperationSystemManagerMock(browser_version),
         chrome_type=chrome_type,